Slim mock_forwarder
