+++ contracts/programs/mock-forwarder/src/lib.rs
@@ -12,27 +12,9 @@
 
 use anchor_lang::prelude::*;
 
-use common::{FORWARDER_METADATA_LENGTH, METADATA_LENGTH, ON_REPORT_DISCRIMINATOR, STATE_VERSION};
+mod internal;
+pub use internal::*;
 
-use events::{
-    ForwarderInitialize, OwnershipAcceptance, OwnershipTransfer, ReportInProgress,
-    ReportProcessed,
-};
-
-use context::*;
-pub use error::*;
-pub use state::{ExecutionState, ForwarderState};
-use utils::{extract_transmission_id, ForwarderReport};
-
-mod common;
-mod context;
-mod error;
-mod events;
-mod state;
-mod utils;
-
-// Placeholder ID (System Program ID). Replace via `anchor keys sync` after
-// `solana-keygen new -o target/deploy/mock_forwarder-keypair.json`.
 declare_id!("7kuEAA3mSC1Tz8gQjnvH7bKFda9xSPRRin9SZbH49cNK");
 
 /// Mock forwarder: relays chainlink reports to a receiver without verifying
@@ -45,61 +27,14 @@ pub mod mock_forwarder {
         hash, instruction::Instruction, program::invoke_signed,
     };
 
-    use crate::utils::extract_raw_report;
-
     use super::*;
 
-    /// Initializes a new mock-forwarder instance and stores data in its state account.
+    /// Creates a new (empty) `ForwarderState` account. The account only exists
+    /// to anchor the `forwarder_authority` PDA — its contents are unused.
     pub fn initialize(ctx: Context<Initialize>) -> Result<()> {
-        let state = &mut ctx.accounts.state;
-        state.version = STATE_VERSION;
-        state.owner = ctx.accounts.owner.key();
-
         emit!(ForwarderInitialize {
-            state: state.key(),
-            owner: ctx.accounts.owner.key(),
+            state: ctx.accounts.state.key(),
         });
-
-        Ok(())
-    }
-
-    /// Step 1 of 2-step ownership process: propose a new owner.
-    pub fn transfer_ownership(
-        ctx: Context<TransferOwnership>,
-        proposed_owner: Pubkey,
-    ) -> Result<()> {
-        let state = &mut ctx.accounts.state;
-        require!(
-            proposed_owner != Pubkey::default()
-                && proposed_owner != state.owner
-                && proposed_owner != state.proposed_owner,
-            ForwarderError::InvalidProposedOwner
-        );
-
-        state.proposed_owner = proposed_owner;
-
-        emit!(OwnershipTransfer {
-            state: state.key(),
-            current_owner: state.owner,
-            proposed_owner: state.proposed_owner
-        });
-
-        Ok(())
-    }
-
-    /// Step 2 of 2-step ownership process: accept ownership.
-    pub fn accept_ownership(ctx: Context<AcceptOwnership>) -> Result<()> {
-        let state = &mut ctx.accounts.state;
-        let state_previous_owner = state.owner;
-        state.owner = state.proposed_owner;
-        state.proposed_owner = Pubkey::default();
-
-        emit!(OwnershipAcceptance {
-            state: state.key(),
-            previous_owner: state_previous_owner,
-            new_owner: state.owner
-        });
-
         Ok(())
     }
